Fix: add soapcpp2_lex.l.patch
On ubuntu 22.04 there is a new Flex/Bison that does not correctly generate files for gsoap.
which, when compiled, produces an error: "multiple definition of `yylval'; soapcpp2-soapcpp2_yacc.o"
This bug has been fixed in gsoap version 2.8.99.
We will update the version of gsoap in the next release, and this patch will be removed!

For more details see: https://sourceforge.net/p/gsoap2/bugs/1269/